YaruCompactLayout: simplify and remove appbar mode
- as discussed in https://github.com/ubuntu/yaru_widgets.dart/issues/211 the NarrowLayout looks different than the WideLayout and also has not much sense for when there are more than 3 or 4 items. The NarrowLayout is also bad for mouse travel on the desktop
- thus this commit removes the layout builder inside compact layout which still leaves the compact layout with three possible widths modes

Closes https://github.com/ubuntu/yaru_widgets.dart/issues/211
